requirements.txt
# Dépendances principales pour le lab IA Pacman
# Compatible avec Python 3.9-3.14 (testé sur 3.14)
numpy>=1.24.0
matplotlib>=3.7.0
jupyter>=1.0.0
pytest>=7.0.0

# Reinforcement Learning
gymnasium>=0.29.0
stable-baselines3>=2.0.0
torch>=2.0.0
pettingzoo>=1.24.0
supersuit>=3.8.0  # wrappers pour PettingZoo

# Interface graphique (remplacement de pygame par pygame-ce pour meilleure compatibilité)
pygame-ce>=2.5.0

# Utilitaires
tkinter  # inclus dans Python standard
